Regression tests for the app
I am using selenium for python to write a functional test.
It is the cleaver way to make an app that works forever.
I do not know if there is some packages for functional tests
in R. I have to search them. For the moment it works very fine.

I have also to think how do unit tests