Change metacommenting, fix the test for default and fail at end
I moved the self.fail at end, the test fail ever, until I finish to write
them all, but If they fail before the end may be a regression.

I used for metacomment `##` - burt the lint does not appretiace them,
I am using instead `# -` and each line of metacomment end with `-`.

At the end. R prints also the indexes for this reason the default of the
page is `'[1] "simple"'`